+++ src/main/java/com/memefest/DataAccess/CategoryFollowerId.java
@@ -1,29 +1,30 @@
 package com.memefest.DataAccess;
 
+
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
 
 @Embeddable
 public class CategoryFollowerId {
   @Column(name = "UserId", nullable = false, updatable = false, insertable = false)
-  private int userId;
+  private Long userId;
   
   @Column(name = "Cat_Id", nullable = false, updatable = false, insertable = false)
-  private int categoryId;
+  private Long categoryId;
   
-  public int getUserId() {
+  public Long getUserId() {
     return this.userId;
   }
   
-  public void setUserId(int userId) {
+  public void setUserId(Long userId) {
     this.userId = userId;
   }
   
-  public int getCat_Id() {
+  public Long getCat_Id() {
     return this.categoryId;
   }
   
-  public void setCat_Id(int categoryId) {
+  public void setCat_Id(Long categoryId) {
     this.categoryId = categoryId;
   }
 }
