+++ src/main/java/com/memefest/DataAccess/PostImageId.java
@@ -1,29 +1,29 @@
 package com.memefest.DataAccess;
 
-    import jakarta.persistence.Column;
+import jakarta.persistence.Column;
     import jakarta.persistence.Embeddable;
 
     @Embeddable
     public class PostImageId {
     @Column(name = "Img_Id", nullable = false, updatable = false, insertable = false)
-    private int imgId;
+    private Long imgId;
     
     @Column(name = "Post_Id", nullable = false, updatable = false, insertable = false)
-    private int postId;
+    private Long postId;
     
-    public int getImg_Id() {
+    public Long getImg_Id() {
         return this.imgId;
     }
     
-    public void setImg_Id(int imgId) {
+    public void setImg_Id(Long imgId) {
         this.imgId = imgId;
     }
     
-    public int getPost_Id() {
+    public Long getPost_Id() {
         return this.postId;
     }
     
-    public void setPost_Id(int postId) {
+    public void setPost_Id(Long postId) {
         this.postId = postId;
     }
     }
