+++ src/main/java/com/memefest/DataAccess/PostReplyId.java
@@ -7,24 +7,24 @@ import jakarta.persistence.Embeddable;
 public class PostReplyId {    
 
   @Column(name = "Post_Id", nullable = false, updatable = false, insertable = false)
-  private int postId;
+  private Long postId;
   
   @Column(name = "Post_Info", nullable = false, updatable = false, insertable = false)
-  private int parentId;
+  private Long parentId;
   
-  public int getPost_Id() {
+  public Long getPost_Id() {
     return this.postId;
   }
   
-  public void setPost_Id(int postId) {
+  public void setPost_Id(Long postId) {
     this.postId = postId;
   }
   
-  public int getPost_Info() {
+  public Long getPost_Info() {
     return this.parentId;
   }
   
-  public void setPost_Info(int parentId) {
+  public void setPost_Info(Long parentId) {
     this.parentId = parentId;
   }
 }
