+++ src/main/java/com/memefest/DataAccess/SubCategoryId.java
@@ -6,24 +6,24 @@ import jakarta.persistence.Embeddable;
 @Embeddable
 public class SubCategoryId {    
   @Column(name = "Cat_Id", nullable = false, updatable = false, insertable = false)
-  private int catId;
+  private Long catId;
   
   @Column(name = "Parent_Id", nullable = false, updatable = false, insertable = false)
-  private int parentId;
+  private Long parentId;
   
-  public int getCat_Id() {
+  public Long getCat_Id() {
     return this.catId;
   }
   
-  public void setCat_Id(int catId) {
+  public void setCat_Id(Long catId) {
     this.catId = catId;
   }
   
-  public int getParent_Id() {
+  public Long getParent_Id() {
     return this.parentId;
   }
   
-  public void setParent_Id(int parentId) {
+  public void setParent_Id(Long parentId) {
     this.parentId = parentId;
   }
 }
