+++ src/main/java/com/memefest/DataAccess/TopicFollower.java
@@ -47,19 +47,19 @@ public class TopicFollower {
     return this.user;
   }
   
-  public void setTopic_Id(int followerId) {
+  public void setTopic_Id(Long followerId) {
     this.id.setTopic_Id(followerId);
   }
   
-  public void setUserId(int userId) {
+  public void setUserId(Long userId) {
     this.id.setUserId(userId);
   }
   
-  public int getTopic_Id() {
+  public Long getTopic_Id() {
     return this.id.getTopic_Id();
   }
   
-  public int getUserId() {
+  public Long getUserId() {
     return this.id.getUserId();
   }
 }
