+++ src/main/java/com/memefest/DataAccess/TopicVideoId.java
@@ -1,6 +1,5 @@
 package com.memefest.DataAccess;
 
-
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
 
@@ -8,24 +7,24 @@ import jakarta.persistence.Embeddable;
 public class TopicVideoId {
 
   @Column(name = "Vid_Id", nullable = false, updatable = false, insertable = false)
-  private int vidId;
+  private Long vidId;
   
   @Column(name = "Topic_Id", nullable = false, updatable = false, insertable = false)
-  private int topicId;
+  private Long topicId;
   
-  public int getVid_Id() {
+  public Long getVid_Id() {
     return this.vidId;
   }
   
-  public void setVid_Id(int vidId) {
+  public void setVid_Id(Long vidId) {
     this.vidId = vidId;
   }
   
-  public int getTopic_Id() {
+  public Long getTopic_Id() {
     return this.topicId;
   }
   
-  public void setTopic_Id(int topicId) {
+  public void setTopic_Id(Long topicId) {
     this.topicId = topicId;
   }
 }
